% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bfmPCA.R
\name{bfmPCA}
\alias{bfmPCA}
\title{sequentially Monitor change with BFAST-PCA}
\usage{
bfmPCA(multibandsarr, history = c("all", "ROC", "BP"), hisweight, timearr,
  pcacomp, moy = 1, myear = 2005, scoreselect, lastordetect = c("last",
  "detect"), sca = F)
}
\arguments{
\item{multibandsarr}{3-d array, space, time and spectral bands}

\item{hisweight}{If True, Use the historical PCA method. If False, use the PCA score method.}

\item{timearr}{The time of the time dimension of the 3-d array}

\item{scoreselect}{if True, select a PC score, specified in pcacomp}

\item{lastordetect}{Use the last observation or detected change date for validation. Useful for PC score method}

\item{sca}{If True, Scale the newly aquired data with the whole time series.}
}
\value{
detected Change time for each pixel
}

