% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/conduct PCA BFAST.R
\name{bfmPCAALL}
\alias{bfmPCAALL}
\title{sequentially monitor change with BFAST-PCA}
\usage{
bfmPCAALL(arr7bands2, history = c("all", "ROC", "BP"), hisweight = T,
  timearr, pcacomp, moy = 1, scoreselect = F, lastordetect = c("last",
  "detect"), sca = F, plot = F)
}
\arguments{
\item{hisweight}{if T, use the historical PCA method. If False, use the PCA score method.}

\item{timearr}{the time of the time dimension of the 4-d array}

\item{sca}{if True scale the newly aquired data with the whole time series.}

\item{arr7bansd2}{4-d array, longitude, latitue, time and spectral bands}
}
\value{
detected change time for each pixel
}

