% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bfmVI.R
\name{bfmVI}
\alias{bfmVI}
\title{sequentially Monitor change with BFAST}
\usage{
bfmVI(multibandsarr, monitoryear = 2005, timearr, moy = 1,
  history = "all", type = "OLS-MOSUM")
}
\arguments{
\item{multibandsarr}{3-d array, space, time and spectral bands}

\item{timearr}{The time of the time dimension of the 3-d array}

\item{moy}{Period for sequential test. moy is 1 means to monitor every 2 years.}
}

