% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/checkseats.R
\name{checkseats}
\alias{checkseats}
\title{check seasonality of time series}
\usage{
checkseats(ts, order, time1)
}
\arguments{
\item{ts}{core data of time series to check}

\item{order}{order of harmonics}

\item{time1}{time of the time series}
}
\value{
rsquare r-sqaure of harmonic fitting
}

