% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mcapply.R
\name{mcapply}
\alias{mcapply}
\title{paralle apply, similar to mclapply and mc.calc}
\usage{
mcapply(arr, margin, fun, mc.cores = 1, ...)
}
\arguments{
\item{mc.cores}{the number of cores to use}
}
\note{
adopt the idea from mc.calc of the bfastSpatial package
}
\author{
Meng Lu
}

