% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mcbfmVI.R
\name{mcbfmVI}
\alias{mcbfmVI}
\title{sequentially Monitor change with BFAST}
\usage{
mcbfmVI(multibandsarr, monitoryear = 2005, timearr, moy = 1,
  history = "all", type = "OLS-MOSUM", mc.cores = 1)
}
\arguments{
\item{multibandsarr}{A 2-d array, with spatial and time as dimensions}

\item{timearr}{The time of the time dimension of the 3-d array}

\item{moy}{Monitoring period for the sequential test. moy is 1 means to monitor every 2 years.}
}
\value{
bfmic Detected change time for each pixel
}

