% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mcwrap.R
\name{mcwrap}
\alias{mcwrap}
\title{wrap for PCA, running different dataset, use multicore}
\usage{
mcwrap(multibandsarr, multibandsarrno, timearr, history, lastordetect,
  hisweight, pcacomp, moy, scoreselect, sca, mc.cores)
}
\arguments{
\item{multibandsarrno}{the second dataset}
}
\value{
time of change for each variable
}

