% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mcwrapvi.R
\name{mcwrapVI}
\alias{mcwrapVI}
\title{wrap for univariable, with multicores}
\usage{
mcwrapVI(multibandsarr, multibandsarrno, timearr, history, pcacomp, moy,
  mc.cores = 1)
}
\value{
time of change for each variable
}

