% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pca_bfm.R
\name{pca_bfm}
\alias{pca_bfm}
\title{Integrate PCA in bfast monitor}
\usage{
pca_bfm(arr, hisweight, pcacomp, myear, history = c("all"), my_dates,
  scoreselect, lastordetect, minumum_observations = 15, sca,
  type = "OLS-MOSUM", moy = 1)
}
\arguments{
\item{arr}{An imput array}
}
\value{
timeofchange detected Change time for each pixel
}

