% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotpcloading.R
\name{plotloading}
\alias{plotloading}
\title{reproduce the plot of pc loading}
\usage{
plotloading(PCfit, varname = "", obsname = "", xaxisname = "", tax = 16,
  plotw = c("time", "bands", "bandsts"), nl = 4, rastertime, addline = 0)
}

