% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plottss2.R
\name{plotts2}
\alias{plotts2}
\title{plot time series produced from return ts, PC score method end a}
\usage{
plotts2(arr, tctl1, timearr, id, nameplot, BTestchangeDate, monitoryear)
}
\arguments{
\item{arr}{array with dimension spectral bands, space, time series}

\item{tctl1}{the index that the landsat 7 starts.}

\item{timearr}{times for the array}

\item{id}{spatial id for the time series plot to return}

\item{nameplot}{the path and name to save the plot}

\item{BTestchangeDate}{validation change date for each location}

\item{monitoryear}{year to start monitoring}
}
\value{
plot time series, as well as the detected and real change time as in the paper. The results is saved.
}

