% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rearrange.R
\name{rearrange_array}
\alias{rearrange_array}
\title{rearrange array from n-d to m-d, with m < n}
\usage{
rearrange_array(original_array, flatten = c(1), position = min(flatten))
}
\arguments{
\item{original_array}{}

\item{flatten}{dimensions to be flattened.}

\item{position}{place /order/position of the new dimension}
}
\value{
rearranged array
}

