% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/returnpc2.R
\name{returnpc2}
\alias{returnpc2}
\title{return PC-greenness and selected PC, for reproducing the results.}
\usage{
returnpc2(inputarr, timearr, loca, preprocess = T, monitoryear)
}
\value{
re book keeping the index of omited value

PChistsauto score of the Historical PCA

PChists2 score of the PC greenness
}

