% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/returnts2.R
\name{returnts2}
\alias{returnts2}
\title{return time series ndmi,ndvi, PCts1, PCts2, PCts3, PCtsauto, PChists1, PChists2, PChists3,PChistsauto, TB, TG, TW. Return also the PCgreeness and selected PC orders, In addition, the removed index is book-keeped}
\usage{
returnts2(inputarr, timearr, tctl1, monitoryear = 2005, loca,
  preprocess = T)
}
\arguments{
\item{inputarr}{3d matrix, [spectral bands, spatial points, time series]}

\item{timearr}{the time of the array}

\item{tctl1}{from which id the Landsat7 is used.}

\item{monitoryear}{the year to start monitoring. By default 2005}

\item{loca}{the spatial id for the time series to be returned}
}
\value{
time stamp to be removed, vegetation index ndmi, ndvi

PC scores from two PCA methods: Historical PC 1-3, auto PC score, PC score 1-3, auto PCscore, time series unscaled.

TB, TG, TW
}

