% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tct.R
\name{tct}
\alias{tct}
\title{tasseled cap transformation on array}
\usage{
tct(arr, l1)
}
\arguments{
\item{arr}{input array}

\item{l1}{from which index the ETM+ is used}
}
\value{
tb cbrightness

tg cgreenness

tw cwetness
}

