% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tctts.R
\name{tctts}
\alias{tctts}
\title{tasseled cap transformation on single location}
\usage{
tctts(arr, l1)
}
\arguments{
\item{arr}{input array}

\item{l1}{from which index the ETM+ is used}
}
\value{
cgreenness

cwetness

cbrightness
}

