% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/uniindex_bfm.R
\name{uniindex_bfm}
\alias{uniindex_bfm}
\title{BFAST sequential monitor on single index}
\usage{
uniindex_bfm(x, moy = 1, myear = 2005, history, my_dates,
  minumum_observations = 15, type = "OLS-MOSUM")
}
\arguments{
\item{x}{input time series coredata}

\item{my_dates}{the time book keeped for the time series}
}
\value{
Detected change time for a time series
}

