% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wrap.R
\name{wrap}
\alias{wrap}
\title{wrap for running different dataset}
\usage{
wrap(multibandsarr, multibandsarrno, timearr, history, lastordetect, hisweight,
  pcacomp, moy, scoreselect, sca)
}
\arguments{
\item{multibandsarrno}{the second dataset}
}

